requirements.txt
langchain~=0.2.12
openai~=1.40.3
SQLAlchemy~=2.0.32
python-dotenv~=1.0.1
pydantic~=2.8.2
fastapi~=0.112.0